batch index and delete

+++ modules/indexer/issues/elastic_search.go
@@ -17,11 +17,11 @@ import (
 )
 
 var (
-	_ Indexer = &ElesticSearchIndexer{}
+	_ Indexer = &ElasticSearchIndexer{}
 )
 
-// ElesticSearchIndexer implements Indexer interface
-type ElesticSearchIndexer struct {
+// ElasticSearchIndexer implements Indexer interface
+type ElasticSearchIndexer struct {
 	client      *elastic.Client
 	indexerName string
 	typeName    string
@@ -35,8 +35,8 @@ func (l elasticLogger) Printf(format string, args ...interface{}) {
 	l.Logger.Log(2, l.Logger.GetLevel(), format, args...)
 }
 
-// NewElesticSearchIndexer creates a new elestic search indexer
-func NewElesticSearchIndexer(url, indexerName string) (*ElesticSearchIndexer, error) {
+// NewElasticSearchIndexer creates a new elasticsearch indexer
+func NewElasticSearchIndexer(url, indexerName string) (*ElasticSearchIndexer, error) {
 	opts := []elastic.ClientOptionFunc{
 		elastic.SetURL(url),
 		elastic.SetSniff(false),
@@ -44,14 +44,14 @@ func NewElesticSearchIndexer(url, indexerName string) (*ElesticSearchIndexer, er
 		elastic.SetGzip(false),
 	}
 
-	logger := log.GetLogger("elastic")
+	logger := elasticLogger{log.GetLogger(log.DEFAULT)}
 
 	if logger.GetLevel() == log.TRACE || logger.GetLevel() == log.DEBUG {
-		opts = append(opts, elastic.SetTraceLog(elasticLogger{logger}))
+		opts = append(opts, elastic.SetTraceLog(logger))
 	} else if logger.GetLevel() == log.ERROR || logger.GetLevel() == log.CRITICAL || logger.GetLevel() == log.FATAL {
-		opts = append(opts, elastic.SetErrorLog(elasticLogger{logger}))
+		opts = append(opts, elastic.SetErrorLog(logger))
 	} else if logger.GetLevel() == log.INFO || logger.GetLevel() == log.WARN {
-		opts = append(opts, elastic.SetInfoLog(elasticLogger{logger}))
+		opts = append(opts, elastic.SetInfoLog(logger))
 	}
 
 	client, err := elastic.NewClient(opts...)
@@ -59,7 +59,7 @@ func NewElesticSearchIndexer(url, indexerName string) (*ElesticSearchIndexer, er
 		return nil, err
 	}
 
-	return &ElesticSearchIndexer{
+	return &ElasticSearchIndexer{
 		client:      client,
 		indexerName: indexerName,
 		typeName:    "indexer_data",
@@ -67,7 +67,7 @@ func NewElesticSearchIndexer(url, indexerName string) (*ElesticSearchIndexer, er
 }
 
 // Init will initial the indexer
-func (b *ElesticSearchIndexer) Init() (bool, error) {
+func (b *ElasticSearchIndexer) Init() (bool, error) {
 	ctx := context.Background()
 	exists, err := b.client.IndexExists(b.indexerName).Do(ctx)
 	if err != nil {
@@ -83,6 +83,10 @@ func (b *ElesticSearchIndexer) Init() (bool, error) {
 			"mappings":{
 				"indexer_data":{
 					"properties":{
+						"id":{
+							"type":"integer",
+							"index": false
+						},
 						"repo_id":{
 							"type":"integer",
 							"index": true
@@ -118,8 +122,11 @@ func (b *ElesticSearchIndexer) Init() (bool, error) {
 }
 
 // Index will save the index data
-func (b *ElesticSearchIndexer) Index(issues []*IndexerData) error {
-	for _, issue := range issues {
+func (b *ElasticSearchIndexer) Index(issues []*IndexerData) error {
+	if len(issues) == 0 {
+		return nil
+	} else if len(issues) == 1 {
+		issue := issues[0]
 		_, err := b.client.Index().
 			Index(b.indexerName).
 			Type(b.typeName).
@@ -132,37 +139,74 @@ func (b *ElesticSearchIndexer) Index(issues []*IndexerData) error {
 				"comments": issue.Comments,
 			}).
 			Do(context.Background())
-		if err != nil {
-			return err
-		}
+		return err
 	}
-	return nil
+
+	reqs := make([]elastic.BulkableRequest, 0)
+	for _, issue := range issues {
+		reqs = append(reqs,
+			elastic.NewBulkIndexRequest().
+				Index(b.indexerName).
+				Type(b.typeName).
+				Id(fmt.Sprintf("%d", issue.ID)).
+				Doc(map[string]interface{}{
+					"id":       issue.ID,
+					"repo_id":  issue.RepoID,
+					"title":    issue.Title,
+					"content":  issue.Content,
+					"comments": issue.Comments,
+				}),
+		)
+	}
+
+	_, err := b.client.Bulk().
+		Index(b.indexerName).
+		Type(b.typeName).
+		Add(reqs...).
+		Do(context.Background())
+	return err
 }
 
 // Delete deletes indexes by ids
-func (b *ElesticSearchIndexer) Delete(ids ...int64) error {
-	for _, id := range ids {
+func (b *ElasticSearchIndexer) Delete(ids ...int64) error {
+	if len(ids) == 0 {
+		return nil
+	} else if len(ids) == 1 {
 		_, err := b.client.Delete().
 			Index(b.indexerName).
 			Type(b.typeName).
-			Id(fmt.Sprintf("%d", id)).
+			Id(fmt.Sprintf("%d", ids[0])).
 			Do(context.Background())
-		if err != nil {
-			return err
-		}
+		return err
 	}
-	return nil
+
+	reqs := make([]elastic.BulkableRequest, 0)
+	for _, id := range ids {
+		reqs = append(reqs,
+			elastic.NewBulkDeleteRequest().
+				Index(b.indexerName).
+				Type(b.typeName).
+				Id(fmt.Sprintf("%d", id)),
+		)
+	}
+
+	_, err := b.client.Bulk().
+		Index(b.indexerName).
+		Type(b.typeName).
+		Add(reqs...).
+		Do(context.Background())
+	return err
 }
 
 // Search searches for issues by given conditions.
 // Returns the matching issue IDs
-func (b *ElesticSearchIndexer) Search(keyword string, repoID int64, limit, start int) (*SearchResult, error) {
+func (b *ElasticSearchIndexer) Search(keyword string, repoID int64, limit, start int) (*SearchResult, error) {
 	kwQuery := elastic.NewMultiMatchQuery(keyword, "title", "content", "comments")
 	query := elastic.NewBoolQuery()
 	query = query.Filter(kwQuery)
 	if repoID > 0 {
 		repoQuery := elastic.NewTermQuery("repo_id", repoID)
-		query = query.Filter(repoQuery)
+		query = query.Must(repoQuery)
 	}
 	searchResult, err := b.client.Search().
 		Index(b.indexerName).
